+++ apps/storybook/.storybook/orderlyProvider.tsx
@@ -17,9 +17,11 @@ import {
 import {
   LedgerWalletAdapter,
   PhantomWalletAdapter,
-  SolflareWalletAdapter
+  SolflareWalletAdapter,
 } from "@solana/wallet-adapter-wallets";
 import config from "../src/config";
+import { Chains } from "@orderly.network/hooks";
+import { NetworkId } from "@orderly.network/types";
 
 const network = WalletAdapterNetwork.Devnet;
 
@@ -55,10 +57,11 @@ const configStore = new CustomConfigStore({
 });
 
 const OrderlyProvider: FC<{ children: ReactNode }> = (props) => {
-  console.log('-- provider', configStore, VITE_ENV);
+  console.log("-- provider", configStore, VITE_ENV);
+
   return (
     <WalletConnectorProvider
-      solanaInitial={{ wallets: wallets}}
+      solanaInitial={{ wallets: wallets }}
       // solanaInitial={{ wallets: wallets, onError: handleSolanaError, network: 'mainnet-beta', mainnetRpc: 'https://svc.blockdaemon.com/solana/mainnet/native?apiKey=zpka_dbb6d1ce22654830860472b76acf15db_62182ef5' }}
     >
       <OrderlyAppProvider
@@ -73,11 +76,17 @@ const OrderlyProvider: FC<{ children: ReactNode }> = (props) => {
           customRestrictedRegions: [],
           contact: { url: "x@orerly.network", text: "x@orerly.network" },
         }}
-
         // overrides={{
         //   tabs: {
         //     variant: "text",
         //   },
+        //   chainSelector: {
+        //     showTestnet: false,
+        //   },
+        // }}
+        // defaultChain={{
+        //   mainnet: { id: 900900900 },
+        //   testnet: { id: 901901901 },
         // }}
       >
         {props.children}