+++ apps/storybook/src/addons/theme_tool/panel.tsx
@@ -1,15 +1,8 @@
-import { useChannel } from "@storybook/manager-api";
 import { AddonPanel } from "@storybook/components";
-import { EVENTS } from "./constants";
 import { ThemeEditor } from "./components/editor";
+import "@orderly.network/ui/dist/styles.css";
 
-export const Panel = (props) => {
-  const emit = useChannel({});
-
-  const handleColorChange = (colorVar: string, value: string) => {
-    emit(EVENTS.CHANGE, { colors: newColors });
-  };
-
+export const Panel = (props: any) => {
   return (
     <AddonPanel {...props}>
       <ThemeEditor />