+++ packages/hooks/src/orderly/useFundingRate.ts
@@ -1,7 +1,11 @@
 import { API } from "@orderly.network/types";
 import { useQuery } from "../useQuery";
 import { useEffect, useMemo, useState } from "react";
-import { Decimal, getTimestamp, timeConvertString } from "@orderly.network/utils";
+import {
+  Decimal,
+  getTimestamp,
+  timeConvertString,
+} from "@orderly.network/utils";
 
 export const useFundingRate = (symbol: string) => {
   if (!symbol) {
@@ -38,7 +42,7 @@ export const useFundingRate = (symbol: string) => {
       }
     }, 1000);
     return () => {
-      clearInterval(timer);
+      clearInterval(timer as unknown as number);
     };
   }, [data]);
 
@@ -51,7 +55,10 @@ export const useFundingRate = (symbol: string) => {
       return null;
     }
 
-    return new Decimal((Number(est_funding_rate) * 100)).toFixed(4, Decimal.ROUND_DOWN);
+    return new Decimal(Number(est_funding_rate) * 100).toFixed(
+      4,
+      Decimal.ROUND_DOWN
+    );
   }, [data]);
 
   return {