+++ packages/cli-codemod/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @orderly.network/codemod
 
+## 1.0.4
+
+### Patch Changes
+
+- fd33fd79c: fix bugs;
+
+## 1.0.4-alpha.0
+
+### Patch Changes
+
+- fix bugs;
+
+## 1.0.3
+
+### Patch Changes
+
+- cca30140b: fix ledger issues
+
+## 1.0.3-alpha.0
+
+### Patch Changes
+
+- fix ledger issues
+
 ## 1.0.2
 
 ### Patch Changes