+++ packages/core/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @orderly.network/core
 
+## 2.0.4
+
+### Patch Changes
+
+- fd33fd79c: fix bugs;
+- Updated dependencies [fd33fd79c]
+  - @orderly.network/types@2.0.4
+
+## 2.0.4-alpha.0
+
+### Patch Changes
+
+- fix bugs;
+- Updated dependencies
+  - @orderly.network/types@2.0.4-alpha.0
+
+## 2.0.3
+
+### Patch Changes
+
+- cca30140b: fix ledger issues
+- Updated dependencies [cca30140b]
+  - @orderly.network/types@2.0.3
+
+## 2.0.3-alpha.0
+
+### Patch Changes
+
+- fix ledger issues
+- Updated dependencies
+  - @orderly.network/types@2.0.3-alpha.0
+
 ## 2.0.2
 
 ### Patch Changes