+++ packages/create-orderly-app/CHANGELOG.md
@@ -1,5 +1,29 @@
 # create-orderly-app
 
+## 1.5.21
+
+### Patch Changes
+
+- fd33fd79c: fix bugs;
+
+## 1.5.21-alpha.0
+
+### Patch Changes
+
+- fix bugs;
+
+## 1.5.20
+
+### Patch Changes
+
+- cca30140b: fix ledger issues
+
+## 1.5.20-alpha.0
+
+### Patch Changes
+
+- fix ledger issues
+
 ## 1.5.19
 
 ### Patch Changes