+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,29 @@
 # tsconfig
 
+## 0.3.20
+
+### Patch Changes
+
+- fd33fd79c: fix bugs;
+
+## 0.3.20-alpha.0
+
+### Patch Changes
+
+- fix bugs;
+
+## 0.3.19
+
+### Patch Changes
+
+- cca30140b: fix ledger issues
+
+## 0.3.19-alpha.0
+
+### Patch Changes
+
+- fix ledger issues
+
 ## 0.3.18
 
 ### Patch Changes